refactor(frontend): migrate layout leaf nodes
