+++ src/frontend/src/features/layout/route-outlet.tsx
@@ -1,35 +1,22 @@
-/**
- * ChatPanel — main content area that renders the active route's component.
- *
- * Uses React Router's `<Outlet />` for content determined by the URL.
- * Wraps the outlet in AnimatePresence for smooth page transitions.
- *
- * The transition key is derived from the active `/app/*` child route
- * so that navigating within a surface (e.g. `/app/workspace` → `/app/volumes`)
- * does NOT re-trigger the page transition animation.
- */
 import { Suspense } from "react";
 import { Outlet, useRouterState } from "@tanstack/react-router";
 import { AnimatePresence, motion, useReducedMotion } from "motion/react";
-import { fades } from "@/lib/utils/motion";
-import { useIsMobile } from "@/hooks/use-is-mobile";
-import { ErrorBoundary } from "@/components/error-boundary";
-import { PageSkeleton } from "@/components/page-skeleton";
 
-export function ShellRouteOutlet() {
+import { ErrorBoundary } from "@/components/patterns/error-boundary";
+import { PageSkeleton } from "@/components/patterns/page-skeleton";
+import { useIsMobile } from "@/hooks/use-is-mobile";
+import { fades } from "@/lib/utils/motion";
+
+export function LayoutRouteOutlet() {
   const routerState = useRouterState();
   const location = routerState.location;
   const isMobile = useIsMobile();
   const prefersReduced = useReducedMotion();
-
-  // Derive the "section" for AnimatePresence key —
-  // only animate when changing top-level sections, not sub-routes
   const section = location.pathname.split("/").filter(Boolean)[1] || "workspace";
 
   return (
-    <div className="flex flex-col h-full w-full bg-background overflow-hidden">
-      {/* Content router */}
-      <div className="flex-1 min-h-0 min-w-0 overflow-hidden">
+    <div className="flex h-full w-full flex-col overflow-hidden bg-background">
+      <div className="min-h-0 min-w-0 flex-1 overflow-hidden">
         <AnimatePresence mode="wait">
           <motion.div
             key={section}
@@ -50,3 +37,5 @@ export function ShellRouteOutlet() {
     </div>
   );
 }
+
+export { LayoutRouteOutlet as ShellRouteOutlet };
